+++ lumin/optimisation/hyper_param.py
@@ -1,4 +1,4 @@
-from typing import Tuple, Dict, List, Optional
+from typing import Tuple, Dict, List, Optional, Any
 from fastprogress import master_bar, progress_bar
 import numpy as np
 from collections import OrderedDict
@@ -12,6 +12,9 @@ from ..nn.data.batch_yielder import BatchYielder
 from ..nn.models.model_builder import ModelBuilder
 from ..nn.models.model import Model
 from ..nn.callbacks.opt_callbacks import LRFinder
+from ..nn.callbacks.cyclic_callbacks import AbsCyclicCallback
+from ..nn.callbacks.model_callbacks import AbsModelCallback
+
 from ..plotting.training import plot_lr_finders
 from ..plotting.plot_settings import PlotSettings
 
@@ -55,7 +58,7 @@ def get_opt_rf_params(x_trn:np.ndarray, y_trn:np.ndarray, x_val:np.ndarray, y_va
 
 def fold_lr_find(fy:FoldYielder, model_builder:ModelBuilder, bs:int,
                  train_on_weights:bool=True, shuffle_fold:bool=True, n_folds:int=-1, lr_bounds:Tuple[float,float]=[1e-5, 10],
-                 plot_settings:PlotSettings=PlotSettings()) -> List[LRFinder]:
+                 callback_args:Dict[str,Dict[str,Any]]={}, plot_settings:PlotSettings=PlotSettings()) -> List[LRFinder]:
     '''Wrapper function for running Smith LR range tests (https://arxiv.org/abs/1803.09820) using folds in FoldYielder'''
     tmr = timeit.default_timer()
     idxs = range(fy.n_folds) if n_folds < 1 else range(min(n_folds, fy.n_folds))
@@ -65,9 +68,18 @@ def fold_lr_find(fy:FoldYielder, model_builder:ModelBuilder, bs:int,
         trn_fold = fy.get_fold(trn_id)
         nb = len(trn_fold['targets'])//bs
         lr_finder = LRFinder(nb=nb, lr_bounds=lr_bounds, model=model)
+        cyclic_callback,callbacks = None,[]
+        for c in callback_args: callbacks.append(c['callback'](**c['kargs']))
+        for c in callbacks:
+            if isinstance(c, AbsCyclicCallback): c.set_nb(nb)
+        for c in callbacks:
+            if isinstance(c, AbsModelCallback): c.set_cyclic_callback(cyclic_callback)
+        for c in callbacks:
+            c.set_model(model)
+            c.on_train_begin()
         lr_finder.on_train_begin()
         batch_yielder = BatchYielder(**fy.get_fold(trn_id), objective=model_builder.objective, bs=bs, use_weights=train_on_weights, shuffle=shuffle_fold)
-        model.fit(batch_yielder, [lr_finder])
+        model.fit(batch_yielder, callbacks+[lr_finder])
         lr_finders.append(lr_finder)
         
     print("LR finder took {:.3f}s ".format(timeit.default_timer()-tmr))